# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AlphaQuant Engine",
    "description": "Backend de dados de mercado (Spot + Futures + multi-exchange) para o AlphaQuant X: indicadores técnicos, estrutura de mercado, Smart Money Concepts, Volume Profile, estatística, derivativos e consenso multi-exchange (preço e estrutura), consumidos sem depender de prints de gráfico.",
    "version": "3.1"
  },
  "servers": [
    {
      "url": "https://alphaquantai-1.onrender.com",
      "description": "Produção (Render)"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "description": "Health check simples -- não toca em nenhuma API externa.",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/snapshot": {
      "get": {
        "summary": "Get Snapshot",
        "description": "Market Snapshot completo: indicadores, estrutura, SMC, volume\nprofile, estatística, derivativos, confluência multi-timeframe e\n(quando habilitado em `config.ENABLE_CROSS_EXCHANGE`) consenso\nmulti-exchange de preço e estrutura. **Endpoint principal.**",
        "operationId": "get_snapshot_snapshot_get",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Par de negociação, ex.: ETHUSDT.",
              "default": "BTCUSDT",
              "title": "Symbol"
            },
            "description": "Par de negociação, ex.: ETHUSDT."
          },
          {
            "name": "timeframes",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Timeframes separados por vírgula, ex.: 15m,1H,4H,1D.",
              "default": "15m,1H,4H,1D",
              "title": "Timeframes"
            },
            "description": "Timeframes separados por vírgula, ex.: 15m,1H,4H,1D."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "additionalProperties": true
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analyze": {
      "get": {
        "summary": "Get Analyze",
        "description": "Análise de um único timeframe -- mantido por compatibilidade (escopo da Fase 1).",
        "operationId": "get_analyze_analyze_get",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Par de negociação, ex.: ETHUSDT.",
              "default": "BTCUSDT",
              "title": "Symbol"
            },
            "description": "Par de negociação, ex.: ETHUSDT."
          },
          {
            "name": "timeframe",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Timeframe único, ex.: 4H.",
              "default": "4H",
              "title": "Timeframe"
            },
            "description": "Timeframe único, ex.: 4H."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "additionalProperties": true
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/scan": {
      "get": {
        "summary": "Get Scan",
        "description": "Varredura multi-símbolo pontual -- ver `scanner/screener.py` para a lógica de classificação.",
        "operationId": "get_scan_scan_get",
        "parameters": [
          {
            "name": "symbols",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Símbolos separados por vírgula (padrão: watchlist de 25 ativos em config.DEFAULT_SCAN_SYMBOLS).",
              "title": "Symbols"
            },
            "description": "Símbolos separados por vírgula (padrão: watchlist de 25 ativos em config.DEFAULT_SCAN_SYMBOLS)."
          },
          {
            "name": "htf",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Timeframe de contexto/tendência.",
              "default": "4H",
              "title": "Htf"
            },
            "description": "Timeframe de contexto/tendência."
          },
          {
            "name": "ltf",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Timeframe de gatilho/execução.",
              "default": "1H",
              "title": "Ltf"
            },
            "description": "Timeframe de gatilho/execução."
          },
          {
            "name": "include_out_of_zone",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Inclui também símbolos sem setup no retorno.",
              "default": false,
              "title": "Include Out Of Zone"
            },
            "description": "Inclui também símbolos sem setup no retorno."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "additionalProperties": true
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/backtest/strategies": {
      "get": {
        "summary": "Get Backtest Strategies",
        "description": "Lista as estratégias registradas e utilizáveis no campo `strategy` de POST /backtest.",
        "operationId": "get_backtest_strategies_backtest_strategies_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestStrategiesResponse"
                }
              }
            }
          }
        }
      }
    },
    "/backtest": {
      "post": {
        "summary": "Post Backtest",
        "description": "Roda backtest bar-a-bar (sem lookahead) de uma estratégia registrada sobre histórico real. Retorna performance (win rate, R médio, profit factor, drawdown) e trades. Erros de dados voltam como HTTP 422 com o motivo, nunca resultado parcial.",
        "operationId": "post_backtest_backtest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BacktestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {},
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BacktestCostModelRequest": {
        "properties": {
          "spread_bps": {
            "type": "number",
            "title": "Spread Bps",
            "default": 0.0
          },
          "slippage_bps": {
            "type": "number",
            "title": "Slippage Bps",
            "default": 0.0
          },
          "commission_bps": {
            "type": "number",
            "title": "Commission Bps",
            "default": 0.0
          }
        },
        "type": "object",
        "title": "BacktestCostModelRequest",
        "description": "Custos de execução (bps). Default é zero em cada campo -- ver `backtest/costs.py`: um backtest sem custo informado é reportado como resultado BRUTO, nunca com fricção \"realista\" inventada."
      },
      "BacktestRequest": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol",
            "description": "Par de negociação, ex.: ETHUSDT. Aceita formato TradingView (ex.: 'BYBIT:ETHUSDT.P')."
          },
          "timeframe": {
            "type": "string",
            "title": "Timeframe",
            "description": "Timeframe dos candles do backtest, ex.: 1H, 4H, 1D."
          },
          "start": {
            "type": "string",
            "format": "date-time",
            "title": "Start",
            "description": "Início do range histórico (ISO 8601)."
          },
          "end": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End",
            "description": "Fim do range histórico (padrão: agora)."
          },
          "strategy": {
            "type": "string",
            "title": "Strategy",
            "description": "Nome da estratégia registrada. Ver GET /backtest/strategies.",
            "default": "sma_cross"
          },
          "strategy_params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Strategy Params",
            "description": "Parâmetros da estratégia (ex.: {\"fast_period\": 10})."
          },
          "cost_model": {
            "$ref": "#/components/schemas/BacktestCostModelRequest"
          },
          "min_candles": {
            "type": "integer",
            "title": "Min Candles",
            "description": "Mínimo de candles exigido no range -- abaixo disso, erro em vez de rodar com amostra insuficiente.",
            "default": 50
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "timeframe",
          "start"
        ],
        "title": "BacktestRequest"
      },
      "BacktestStrategiesResponse": {
        "properties": {
          "strategies": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Strategies"
          }
        },
        "type": "object",
        "required": [
          "strategies"
        ],
        "title": "BacktestStrategiesResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "HealthResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}